feat(F-579): clean lib folder
