+++ frontend/src/app/[locale]/(auth)/confirm/components/ConfirmationForm.tsx
@@ -4,9 +4,9 @@ import { Button } from '@/components/ui/Button';
 import { Card, CardContent, CardFooter } from '@/components/ui/Card';
 import { Form, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/Form';
 import Input from '@/components/ui/Input';
-import { handlePasteEvent } from '@/lib/handlePaste';
+import { handlePasteEvent } from '@/lib/handlers/handlePaste';
 import { createClient } from '@/lib/supabase/client';
-import { showErrorToast } from '@/lib/toast';
+import { showErrorToast } from '@/lib/component-utils/toast';
 import { api } from '@/services/ApiClient';
 import { authService } from '@/services/AuthService';
 import { zodResolver } from '@hookform/resolvers/zod';
