+++ frontend/src/app/[locale]/(main)/settings/components/Settings.tsx
@@ -15,9 +15,9 @@ import UserConfidenceFields from '@/components/common/UserConfidenceFields';
 import { DeleteUserHandler } from '@/components/common/DeleteUserHandler';
 import { UserProfileService } from '@/services/UserProfileService';
 import { UserPreference } from '@/interfaces/models/UserInputFields';
-import { PrimaryGoals, UserRoles } from '@/lib/utils';
+import { PrimaryGoals, UserRoles } from '@/lib/common';
 import { UserProfile } from '@/interfaces/models/UserProfile';
-import { showErrorToast, showSuccessToast } from '@/lib/toast';
+import { showErrorToast, showSuccessToast } from '@/lib/component-utils/toast';
 import JSZip from 'jszip';
 import { api } from '@/services/ApiClient';
 import UserPreferences from './UserPreferences';
