+++ frontend/src/lib/handlers/handlePaste.ts
